feat: add route for generating random data file
This commit adds a new route in the main.go file to handle requests for generating random data files. The route accepts a parameter for the file size in megabytes and creates a file with approximately that size. If the file already exists, it is returned as a download. Otherwise, a new file is created and filled with random data.